enum 클래스 속성 반환 메서드 추가 및 코드 개선

+++ src/main/java/com/example/customermanageproejctproduct/dto/AddProductRequest.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class AddProductRequest {
     private String productName;
     private ProductCategory category;
-    private int price;
-    private int stock;
+    private long price;
+    private long stock;
     private ProductStatus status;
 }
